Restructure useing extensions
