+++ analysis_options.yaml
@@ -1,1 +1,5 @@
 include: package:lint/analysis_options_package.yaml
+
+analyzer:
+  strong-mode:
+    implicit-casts: false
