+++ pubspec.yaml
@@ -5,7 +5,7 @@ homepage: https://github.com/passsy/deep_pick
 author: Pascal Welsch <pascal.welsch@gmail.com>
 
 environment:
-  sdk: '>=2.4.0 <3.0.0'
+  sdk: '>=2.6.0-dev.8.0 <3.0.0'
 
 dev_dependencies:
   test: ^1.6.0
